+++ corehq/apps/formplayer_api/management/commands/clear_formplayer_dbs.py
@@ -9,9 +9,9 @@ from django.core.management.base import BaseCommand
 from corehq.apps.formplayer_api.management.commands.prime_formplayer_restores import (
     _get_user_rows, _get_users_from_csv
 )
+from corehq.apps.formplayer_api.tasks import get_prime_restore_user_params, clear_formplayer_db_for_user
 from corehq.apps.users.models import CouchUser
 from corehq.util.argparse_types import validate_range
-from custom.formplayer.restore_priming import get_prime_restore_user_params, clear_formplayer_db_for_user
 
 
 class Command(BaseCommand):
