+++ corehq/apps/formplayer_api/management/commands/prime_formplayer_restores.py
@@ -7,13 +7,13 @@ from datetime import datetime
 from dateutil.relativedelta import relativedelta
 from django.core.management.base import BaseCommand
 
-from corehq.apps.users.models import CouchUser
-from corehq.util.argparse_types import validate_range
-from custom.formplayer.restore_priming import (
+from corehq.apps.formplayer_api.tasks import (
     get_users_for_priming,
     get_prime_restore_user_params,
     prime_formplayer_db_for_user,
 )
+from corehq.apps.users.models import CouchUser
+from corehq.util.argparse_types import validate_range
 
 
 class Command(BaseCommand):
